multithreading

+++ src/test/kotlin/rasteplads/api/eventmesh_test/BuilderTest.kt
@@ -173,52 +173,52 @@ class BuilderTest {
         val f = correct()
         var name = "msgDelete"
         var default =
-            getValueFromClass<EventMesh<Int, Byte, MessageCache<Int>>, UInt>(f.build(), name)
+            getValueFromClass<EventMesh<Int, Byte, MessageCache<Int>>, Long>(f.build(), name)
         var modded =
-            getValueFromClass<EventMesh<Int, Byte, MessageCache<Int>>, UInt>(
-                f.withMsgCacheDelete(default + 10u).build(), name)
+            getValueFromClass<EventMesh<Int, Byte, MessageCache<Int>>, Long>(
+                f.withMsgCacheDelete(default + 10).build(), name)
         assertNotEquals(modded, default)
 
         name = "msgTTL"
-        default = getValueFromClass<EventMesh<Int, Byte, MessageCache<Int>>, UInt>(f.build(), name)
+        default = getValueFromClass<EventMesh<Int, Byte, MessageCache<Int>>, Long>(f.build(), name)
         modded =
-            getValueFromClass<EventMesh<Int, Byte, MessageCache<Int>>, UInt>(
-                f.withMsgTTL(default + 10u).build(), name)
+            getValueFromClass<EventMesh<Int, Byte, MessageCache<Int>>, Long>(
+                f.withMsgTTL(default + 10).build(), name)
         assertNotEquals(modded, default)
 
         name = "msgSendInterval"
-        default = getValueFromClass<EventMesh<Int, Byte, MessageCache<Int>>, UInt>(f.build(), name)
+        default = getValueFromClass<EventMesh<Int, Byte, MessageCache<Int>>, Long>(f.build(), name)
         modded =
-            getValueFromClass<EventMesh<Int, Byte, MessageCache<Int>>, UInt>(
-                f.withMsgSendInterval(default + 10u).build(), name)
+            getValueFromClass<EventMesh<Int, Byte, MessageCache<Int>>, Long>(
+                f.withMsgSendInterval(default + 10).build(), name)
         assertNotEquals(modded, default)
 
         name = "msgSendDuration"
-        default = getValueFromClass<EventMesh<Int, Byte, MessageCache<Int>>, UInt>(f.build(), name)
+        default = getValueFromClass<EventMesh<Int, Byte, MessageCache<Int>>, Long>(f.build(), name)
         modded =
-            getValueFromClass<EventMesh<Int, Byte, MessageCache<Int>>, UInt>(
-                f.withMsgSendDuration(default + 10u).build(), name)
+            getValueFromClass<EventMesh<Int, Byte, MessageCache<Int>>, Long>(
+                f.withMsgSendDuration(default + 10).build(), name)
         assertNotEquals(modded, default)
 
         name = "msgScanInterval"
-        default = getValueFromClass<EventMesh<Int, Byte, MessageCache<Int>>, UInt>(f.build(), name)
+        default = getValueFromClass<EventMesh<Int, Byte, MessageCache<Int>>, Long>(f.build(), name)
         modded =
-            getValueFromClass<EventMesh<Int, Byte, MessageCache<Int>>, UInt>(
-                f.withMsgScanInterval(default + 10u).build(), name)
+            getValueFromClass<EventMesh<Int, Byte, MessageCache<Int>>, Long>(
+                f.withMsgScanInterval(default + 10).build(), name)
         assertNotEquals(modded, default)
 
         name = "msgScanDuration"
-        default = getValueFromClass<EventMesh<Int, Byte, MessageCache<Int>>, UInt>(f.build(), name)
+        default = getValueFromClass<EventMesh<Int, Byte, MessageCache<Int>>, Long>(f.build(), name)
         modded =
-            getValueFromClass<EventMesh<Int, Byte, MessageCache<Int>>, UInt>(
-                f.withMsgScanDuration(default + 10u).build(), name)
+            getValueFromClass<EventMesh<Int, Byte, MessageCache<Int>>, Long>(
+                f.withMsgScanDuration(default + 10).build(), name)
         assertNotEquals(modded, default)
 
         name = "msgCacheLimit"
-        default = getValueFromClass<EventMesh<Int, Byte, MessageCache<Int>>, UInt>(f.build(), name)
+        default = getValueFromClass<EventMesh<Int, Byte, MessageCache<Int>>, Long>(f.build(), name)
         modded =
-            getValueFromClass<EventMesh<Int, Byte, MessageCache<Int>>, UInt>(
-                f.withMsgCacheLimit(default + 10u).build(), name)
+            getValueFromClass<EventMesh<Int, Byte, MessageCache<Int>>, Long>(
+                f.withMsgCacheLimit(default + 10).build(), name)
         assertNotEquals(modded, default)
     }
